# docs/api-spec/loyalty-rewards-service-openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: loyalty@probestack.io
    name: Growth Engineering Team
  description: |
    Manages loyalty point balances, tier progression, and reward redemptions.
    This API is used as the source specification to automatically generate
    an API Gateway proxy (Kong or Apigee) with routing, authentication,
    rate limiting, and caching policies.
  title: Loyalty Rewards Service
  version: 1.0.0
  x-proxy-config:
    upstream-url: https://loyalty-service.probestack.io
    service-name: loyalty-rewards-gateway
    rate-limit: 250/minute
    cors-enabled: true
    retries: 2
servers:
- description: Production Gateway
  url: https://gateway.probestack.io/loyalty/v1
- description: Local Gateway
  url: http://localhost:8000/loyalty/v1
security:
- ApiKeyAuth: []
...
paths:
  /accounts/{memberId}:
    get:
      description: "Retrieves the point balance, current tier, and progress towards\
        \ the next tier for a member."
      operationId: getLoyaltyAccount
      parameters:
      - description: Unique loyalty member identifier
        explode: false
        in: path
        name: memberId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccount"
          description: Loyalty account fetched successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Get loyalty account
      tags:
      - Accounts
      x-proxy-route:
        methods:
        - GET
        plugins:
        - name: cors
        - name: rate-limiting
          config:
            minute: 250
        - name: proxy-cache
          config:
            cache_ttl: 30
      x-accepts: application/json
      x-tags:
      - tag: Accounts
    put:
      description: Updates the member's communication and auto-redeem preferences.
        Point balance cannot be set directly.
      operationId: updateLoyaltyAccount
      parameters:
      - explode: false
        in: path
        name: memberId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccount"
          description: Account updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request payload or parameters.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Update account preferences
      tags:
      - Accounts
      x-proxy-route:
        methods:
        - PUT
        plugins:
        - name: cors
        - name: rate-limiting
          config:
            minute: 100
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Accounts
...
components:
...
  schemas:
    LoyaltyAccount:
      example:
        tier: GOLD
        emailUpdates: true
        pointsToNextTier: 2550
        autoRedeem: false
        lifetimePoints: 48200
        pointBalance: 12450
        memberId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        enrolledAt: 2000-01-23T04:56:07.000+00:00
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        memberId:
          format: uuid
          type: string
        pointBalance:
          example: 12450
          type: integer
        lifetimePoints:
          example: 48200
          type: integer
        tier:
          enum:
          - BRONZE
          - SILVER
          - GOLD
          - PLATINUM
          example: GOLD
          type: string
        pointsToNextTier:
          example: 2550
          type: integer
        autoRedeem:
          example: false
          type: boolean
        emailUpdates:
          example: true
          type: boolean
        enrolledAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - memberId
      - pointBalance
      - tier
      type: object
    UpdateAccountRequest:
      example:
        emailUpdates: false
        autoRedeem: true
      properties:
        autoRedeem:
          example: true
          type: boolean
        emailUpdates:
          example: false
          type: boolean
      type: object
...
    ErrorResponse:
      properties:
        status:
          example: 400
          type: integer
        errorCode:
          example: INSUFFICIENT_POINTS
          type: string
        message:
          example: Member does not have enough points for this reward
          type: string
        timestamp:
          format: date-time
          type: string
        path:
          example: /loyalty/v1/rewards
          type: string
      required:
      - message
      - status
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: API Key required for all gateway requests.
      in: header
      name: X-API-Key
      type: apiKey