# src/main/java/com/probestack/forgestudio/design/api/AccountsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.LoyaltyAccount;
import java.util.UUID;
import com.probestack.forgestudio.design.model.UpdateAccountRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-13T10:36:55.986768776Z[GMT]")@Validated
@Tag(name = "Accounts", description = "Loyalty account balance and tier")
public interface AccountsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /accounts/{memberId} : Get loyalty account
     * Retrieves the point balance, current tier, and progress towards the next tier for a member.
     *
     * @param memberId Unique loyalty member identifier (required)
     * @return Loyalty account fetched successfully (status code 200)
     *         or Missing or invalid API Key. (status code 401)
     *         or Resource not found. (status code 404)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "getLoyaltyAccount",
        summary = "Get loyalty account",
        description = "Retrieves the point balance, current tier, and progress towards the next tier for a member.",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Loyalty account fetched successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoyaltyAccount.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accounts/{memberId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<LoyaltyAccount> getLoyaltyAccount(
        @Parameter(name = "memberId", description = "Unique loyalty member identifier", required = true, in = ParameterIn.PATH) @PathVariable("memberId") UUID memberId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"tier\" : \"GOLD\", \"emailUpdates\" : true, \"pointsToNextTier\" : 2550, \"autoRedeem\" : false, \"lifetimePoints\" : 48200, \"pointBalance\" : 12450, \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"enrolledAt\" : \"2000-01-23T04:56:07.000+00:00\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * PUT /accounts/{memberId} : Update account preferences
     * Updates the member&#39;s communication and auto-redeem preferences. Point balance cannot be set directly.
     *
     * @param memberId  (required)
     * @param updateAccountRequest  (required)
     * @return Account updated successfully (status code 200)
     *         or Invalid request payload or parameters. (status code 400)
     *         or Missing or invalid API Key. (status code 401)
     *         or Resource not found. (status code 404)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "updateLoyaltyAccount",
        summary = "Update account preferences",
        description = "Updates the member's communication and auto-redeem preferences. Point balance cannot be set directly.",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Account updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoyaltyAccount.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/accounts/{memberId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<LoyaltyAccount> updateLoyaltyAccount(
        @Parameter(name = "memberId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("memberId") UUID memberId,
        @Parameter(name = "UpdateAccountRequest", description = "", required = true) @Valid @RequestBody UpdateAccountRequest updateAccountRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"tier\" : \"GOLD\", \"emailUpdates\" : true, \"pointsToNextTier\" : 2550, \"autoRedeem\" : false, \"lifetimePoints\" : 48200, \"pointBalance\" : 12450, \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"enrolledAt\" : \"2000-01-23T04:56:07.000+00:00\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}